services/content-ideas-agent/requirements.txt
anthropic>=0.40.0
supabase>=2.9.0
langgraph>=0.2.50
langchain-core>=0.3.20
feedparser>=6.0.11
python-telegram-bot>=21.6
APScheduler>=3.10.4
structlog>=24.4.0
PyYAML>=6.0.2
python-dotenv>=1.0.1
httpx>=0.27.2
tenacity>=9.0.0
